more app.py updates
